+++ packages/vue/src/number-input/number-input-context.ts
@@ -1,9 +1,7 @@
-import { type connect } from '@zag-js/number-input'
-import { type ComputedRef } from 'vue'
 import { createContext } from '../context'
 import { type UseNumberInputReturn } from './use-number-input'
 
-export const [NumberInputProvider, useNumberInputContext] =
-  createContext<ComputedRef<ReturnType<typeof connect>>>('NumberInputContext')
-
 export type NumberInputContext = UseNumberInputReturn
+
+export const [NumberInputProvider, useNumberInputContext] =
+  createContext<NumberInputContext>('NumberInputContext')