+++ packages/vue/src/number-input/number-input-control.tsx
@@ -1,11 +1,11 @@
 import { defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
-import { getValidChildren, type ComponentWithProps } from '../utils'
+import { getValidChildren } from '../utils'
 import { useNumberInputContext } from './number-input-context'
 
 export type NumberInputControlProps = HTMLArkProps<'div'>
 
-export const NumberInputControl: ComponentWithProps<NumberInputControlProps> = defineComponent({
+export const NumberInputControl = defineComponent({
   name: 'NumberInputControl',
   setup(_, { slots, attrs }) {
     const api = useNumberInputContext()