+++ packages/vue/src/number-input/number-input-field.tsx
@@ -1,11 +1,10 @@
 import { computed, defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
-import type { ComponentWithProps } from '../utils'
 import { useNumberInputContext } from './number-input-context'
 
 export type NumberInputInputProps = HTMLArkProps<'input'>
 
-export const NumberInputInput: ComponentWithProps<NumberInputInputProps> = defineComponent({
+export const NumberInputInput = defineComponent({
   name: 'NumberInputInput',
   setup(_, { attrs }) {
     const api = useNumberInputContext()